fix: chatbot does not greet on slow 4 devices (#193)
# Type of Change

- [ ] BREAKING CHANGE (change that would cause existing functionality to
not work as expected)
- [ ] Feature (change that implements a new feature)
- [x] Bug Fix (change that fixed a bug)
- [x] Improvements (change that improves existing features or
performance)
- [ ] Documentation (change in documents)
- [ ] Chore (formatting, refactoring, styling or other operations
commit)

###

# Description

The chatbot doesn't immediately greet on first time visits due to the
Link tag provided by react router that causes the page to load
immediately. This doesn't give enough time for the bot to post a chat
and greet the user before the user opened the chat head. This leads us
to require the users to reload first. This is now fixed by adding some
checking to the chat head if the length of the conversation is still 0,
then don't show yet. Also replace the Link tag provided by react router
into a default anchor.

Co-authored-by: MoonbamiOfficial <141120384+MoonbamiOfficial@users.noreply.github.com>