fix: chat head checking to make use of chat skeleton screens (#224)
# Type of Change

<!-- Check what's necessary or check all if applicable. -->

- [ ] BREAKING CHANGE (change that would cause existing functionality to
not work as expected)
- [ ] Feature (change that implements a new feature)
- [x] Bug Fix (change that fixed a bug)
- [x] Improvements (change that improves existing features or
performance)
- [ ] Documentation (change in documents)
- [x] Chore (formatting, refactoring, styling or other operations
commit)

###

# Description

Since before, we're checking if the chatbot is at the latest chat. The
chat skeleton screen is being useless. Updated the checking in the chat
head so it will appear immediately and while the conversation.length is
still 0, chat skeleton screen will handle the supposed ui.

---------

Co-authored-by: MoonbamiOfficial <141120384+MoonbamiOfficial@users.noreply.github.com>